refactor(node): Prepare NPM listing helpers for reuse
Move the ModuleInfo extension functions to ModuleInfo.kt and make them
internal, add a source parameter to extractNpmIssues(), and make the
package manager type of NpmDependencyHandler configurable, to prepare
for adding Bun support.
