+++ plugins/package-managers/node/src/main/kotlin/npm/ModuleInfo.kt
@@ -19,9 +19,13 @@
 
 package org.ossreviewtoolkit.plugins.packagemanagers.node.npm
 
+import java.util.LinkedList
+
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 import kotlinx.serialization.json.Json
+
+import org.ossreviewtoolkit.plugins.packagemanagers.node.Scope
 
 private val JSON = Json { ignoreUnknownKeys = true }
 
@@ -61,3 +65,62 @@ internal data class ModuleInfo(
     /** The URI from where the package was resolved. Starts with "file:" for local packages. */
     val resolved: String? = null
 )
+
+internal fun ModuleInfo.getAllPackageNodeModuleIds(scopes: Set<Scope>): Set<String> =
+    buildSet {
+        val queue = scopes.flatMapTo(LinkedList()) { getScopeDependencies(it) }
+
+        while (queue.isNotEmpty()) {
+            val info = queue.removeFirst()
+
+            @Suppress("ComplexCondition")
+            if (!info.isProject && info.isInstalled && !info.name.isNullOrBlank() && !info.version.isNullOrBlank()) {
+                add("${info.name}@${info.version}")
+            }
+
+            scopes.flatMapTo(queue) { info.getScopeDependencies(it) }
+        }
+    }
+
+internal fun ModuleInfo.getScopeDependencies(scope: Scope) =
+    when (scope) {
+        Scope.DEPENDENCIES -> dependencies.values.filter { !it.dev }
+        Scope.DEV_DEPENDENCIES -> dependencies.values.filter { it.dev && !it.optional }
+    }
+
+internal fun ModuleInfo.undoDeduplication(): ModuleInfo {
+    val replacements = getNonDeduplicatedModuleInfosForId()
+
+    fun ModuleInfo.undoDeduplicationRec(ancestorsIds: Set<String> = emptySet()): ModuleInfo {
+        val dependencyAncestorIds = ancestorsIds + setOfNotNull(id)
+        val replacement = replacements[id] ?: this
+
+        return replacement.copy(
+            dependencies = replacement.dependencies
+                .filter { it.value.id !in dependencyAncestorIds } // break cycles.
+                .mapValues { it.value.undoDeduplicationRec(dependencyAncestorIds) }
+        )
+    }
+
+    return undoDeduplicationRec()
+}
+
+internal fun ModuleInfo.filterInstalled(): ModuleInfo =
+    copy(dependencies = dependencies.filter { it.value.isInstalled })
+
+private fun ModuleInfo.getNonDeduplicatedModuleInfosForId(): Map<String, ModuleInfo> {
+    val queue = LinkedList<ModuleInfo>().apply { add(this@getNonDeduplicatedModuleInfosForId) }
+    val result = mutableMapOf<String, ModuleInfo>()
+
+    while (queue.isNotEmpty()) {
+        val info = queue.removeFirst()
+
+        if (info.id != null && info.dependencyConstraints.keys.subtract(info.dependencies.keys).isEmpty()) {
+            result[info.id] = info
+        }
+
+        queue += info.dependencies.values
+    }
+
+    return result
+}
